Create Mutations.js
My first attempt (when trying to stick with RegEx) wasn't working -- I spent an afternoon searching for a way to match some but not all but am unsure if this is the best method due to the pattern-finding (looking-for?) nature of RegEx. Am shelving this for now and will try a different method.